requirements.txt
Pillow
pyautogui
numpy